source/backend/requirements.txt
Flask==2.0.2
requests==2.27.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
flask-swagger-ui==3.36.0
vowpalwabbit==8.11.0
pandas==1.3.5
matplotlib==3.5.1
pytest==6.2.5
pycamunda==0.5.0
scipy==1.7.3